authorization

+++ routers/posts.py
@@ -5,6 +5,7 @@ from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
 
+from auth import CurrentUser
 import models
 from database import get_db
 from schemas import PostCreate, PostResponse, PostUpdate
@@ -14,7 +15,7 @@ router = APIRouter()
 
 @router.get('', response_model=list[PostResponse])
 async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
-    result = db.execute(select(models.Post).options(selectinload(models.Post.author)))
+    result = await db.execute(select(models.Post).options(selectinload(models.Post.author)))
     posts = result.scalars().all()
     return posts
 
@@ -24,17 +25,14 @@ async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
     response_model=PostResponse,
     status_code=status.HTTP_201_CREATED,
     )
-async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def create_post(post: PostCreate,
+                    current_user: CurrentUser,
+                    db: Annotated[AsyncSession, Depends(get_db)]):
 
-    result = await db.execute(select(models.User).where(models.User.id==post.user_id))
-    user = result.scalars().first()
-    if not user:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
-    
     new_post = models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id
+        user_id=current_user.id,
     )
     
     db.add(new_post)
@@ -45,7 +43,7 @@ async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_
 
 @router.get('/{post_id}', response_model=PostResponse)
 async def get_post_using_postid(post_id : int, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = db.execute(select(models.Post)
+    result = await db.execute(select(models.Post)
                         .options(selectinload(models.Post.author))
                         .where(models.Post.id==post_id))
     post = result.scalars().first()
@@ -55,35 +53,41 @@ async def get_post_using_postid(post_id : int, db: Annotated[AsyncSession, Depen
     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
 
 @router.put('/{post_id}', response_model=PostResponse)
-async def update_post_full(post_id: int, post_data: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_full(post_id: int,
+                            post_data: PostCreate,
+                            current_user: CurrentUser,
+                            db: Annotated[AsyncSession, Depends(get_db)]):
+    
     result = await db.execute(select(models.Post).options(selectinload(models.Post.author)).where(models.Post.id==post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
     
-    if post.user_id != post_data.user_id:
-        result = await db.execute(select(models.User).where(models.User.id==post_data.user_id))
-        user = result.scalars().first()
-        if not user:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You are not authorized to update this post")
         
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
 
     await db.commit()
-    await db.refresh(post, attribute_names=["Author"])
-
+    await db.refresh(post, attribute_names=["author"])
     return post
 
 @router.patch('/{post_id}', response_model=PostResponse)
-async def update_post_partial(post_id: int, post_data: PostUpdate, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = db.execute(select(models.Post)
+async def update_post_partial(post_id: int,
+                            post_data: PostUpdate,
+                            current_user: CurrentUser,
+                            db: Annotated[AsyncSession, Depends(get_db)]):
+    
+    result = await db.execute(select(models.Post)
                         .options(selectinload(models.Post.author))
                         .where(models.Post.id==post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
+
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You are not authorized to update this post")
     
     # important *
     update_data = post_data.model_dump(exclude_unset=True) # -> dict
@@ -91,11 +95,13 @@ async def update_post_partial(post_id: int, post_data: PostUpdate, db: Annotated
         setattr(post, field, value)
 
     await db.commit()
-    await db.refresh(post, attribute_names=["Author"])
+    await db.refresh(post, attribute_names=["author"])
     return post
 
 @router.delete('/{post_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id : int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(post_id : int,
+                    current_user: CurrentUser,
+                    db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(select(models.Post)
                               .options(selectinload(models.Post.author))
                               .where(models.Post.id==post_id))
@@ -103,5 +109,8 @@ async def delete_post(post_id : int, db: Annotated[AsyncSession, Depends(get_db)
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
     
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You are not authorized to delete this post")
+
     await db.delete(post)
     await db.commit()
